+++ src/hooks/prerun/check-dependencies.ts
@@ -1,8 +1,8 @@
 /* jscpd:ignore-start */
 
 import * as os from "os";
-import { checkSfdxPlugin, git, uxLog, execCommand, isCI, execSfdxJson, checkAppDependency, isGitRepo } from "../../common/utils";
-import { getConfig, setConfig } from "../../config";
+import { checkSfdxPlugin, git, uxLog, isCI, checkAppDependency, isGitRepo } from "../../common/utils";
+import { getConfig } from "../../config";
 
 export const hook = async (options: any) => {
   // Skip hooks from other commands than hardis commands
@@ -12,28 +12,6 @@ export const hook = async (options: any) => {
   }
   if (commandId.startsWith("hardis:doc") || commandId.startsWith("hardis:org:files") || commandId.startsWith("hardis:org:data")) {
     return;
-  }
-
-  // Set only once restDeploy=false to improve performances
-  const config = await getConfig("user");
-  if (commandId.includes("deploy") || commandId.includes("scratch:create")) {
-    if (!(config.forceRestDeploy === true)) {
-      if (config.restDeployDisabled !== true && !commandId.includes("configure")) {
-        execSfdxJson("sfdx config:get restDeploy", {
-          output: false,
-          fail: true,
-          spinner: false,
-        }).then(async (res) => {
-          if (!(res && res.result && res.result[0] && res.result[0].value === "false")) {
-            await execCommand("sfdx config:set restDeploy=false --global", {
-              output: false,
-              fail: true,
-            });
-          }
-          await setConfig("user", { restDeployDisabled: true });
-        });
-      }
-    }
   }
 
   /* jscpd:ignore-end */